web: add secret values redaction to object export
